Live blog posting SEO updates
